Move to Cosmos SDK collections  (#284)
* Move collateral to cosmos collections

* Move Jailed over

* Try IndexedMap

* Participants

* Fix unit tests - standardize on "gonka" prefix

* Update RandomSeeds

* Now poc batches/validation

* Now ActiveParticipants

* Revert "Now ActiveParticipants"

This reverts commit d1566649c9b0bea50b89cd0d5a367852895f4bd9.

* Model Pricing

* Move keys to keys.

* Models

* Timeouts

* No longer needed file

* Inference validation details

* Unit of compute

* Fix merge
Still something weird in SettleAccounts... want to run IntegrationTests

* Yikes! Major mistake in prefixes

* Epochs

* Epoch Group Validation

* Inferences

* Settle Amount

* Top Miners

* Partial Upgrades

* Fix build

* Epoch PerformanceSummaries

* Delete key file

* streamvesting module

* Save the schema

+++ inference-chain/x/inference/module/chainvalidation.go
@@ -307,7 +307,7 @@ func (am AppModule) ComputeNewWeights(ctx context.Context, upcomingEpoch types.E
 		}
 		participants[participantAddress] = participant
 
-		seed, found := am.keeper.GetRandomSeed(ctx, int64(upcomingEpoch.Index), participantAddress)
+		seed, found := am.keeper.GetRandomSeed(ctx, upcomingEpoch.Index, participantAddress)
 		if !found {
 			am.LogError("ComputeNewWeights: Participant didn't submit the seed for the upcoming epoch", types.PoC,
 				"upcomingEpoch.Index", upcomingEpoch.Index,
@@ -321,7 +321,7 @@ func (am AppModule) ComputeNewWeights(ctx context.Context, upcomingEpoch types.E
 	// STEP 3: Add seeds for preserved participants if they have submitted seeds
 	for _, preservedParticipant := range preservedParticipants {
 		participantAddress := preservedParticipant.Index
-		if seed, found := am.keeper.GetRandomSeed(ctx, int64(upcomingEpoch.Index), participantAddress); found {
+		if seed, found := am.keeper.GetRandomSeed(ctx, upcomingEpoch.Index, participantAddress); found {
 			preservedParticipant.Seed = &seed
 			seeds[participantAddress] = seed
 			am.LogInfo("ComputeNewWeights: Added seed for preserved participant", types.PoC,
